Add changes to fix interrupt flow
